+++ packages/opentelemetry-node/CHANGELOG.md
@@ -2,8 +2,14 @@
 
 ## Unreleased
 
+- feat: Add `@opentelemetry/instrumentation-mysql` to the default set
+  of instrumentations. See <https://github.com/open-telemetry/opentelemetry-js-contrib/blob/main/plugins/node/opentelemetry-instrumentation-mysql#readme>
 - feat: Add `@opentelemetry/instrumentation-mysql2` to the default set
   of instrumentations. See <https://github.com/open-telemetry/opentelemetry-js-contrib/blob/main/plugins/node/opentelemetry-instrumentation-mysql2#readme>
+- feat: Add `@opentelemetry/instrumentation-cassandra-driver` to the default set
+  of instrumentations. See <https://github.com/open-telemetry/opentelemetry-js-contrib/blob/main/plugins/node/opentelemetry-instrumentation-cassandra#readme>
+- test: Test that the native instrumentation in `@elastic/elasticsearch@8.15.0` and later works.
+
 
 ## v0.6.0